memperbaiki rute pada edit pet
