+++ app/src/main/java/com/varsha/pawpals/ui/presentation/schedule/editPlan/DailyPlanItem.kt
@@ -35,7 +35,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.varsha.pawpals.utils.AlarmData
+import com.varsha.pawpals.model.AlarmData
+
 
 @Composable
 fun DailyPlanItem(
